Try deploying bolt app w/ Serverless
So far it's not working, the `serverless deploy` is failing when it
tries to push to the s3 bucket that it creates...
I've tried looking around in the documentation but it seems like
configuring serverless from a YAML file is outdated anyways.

Anyways I'm committing this for now but I'm gonna backtrack and just go
with a docker-based deployment strategy instead.